+++ src/DependencyInjection/SentryExtension.php
@@ -14,6 +14,7 @@ use Sentry\Integration\IntegrationInterface;
 use Sentry\Integration\RequestFetcherInterface;
 use Sentry\Integration\RequestIntegration;
 use Sentry\Options;
+use Sentry\SentryBundle\EventListener\ConsoleListener;
 use Sentry\SentryBundle\EventListener\ErrorListener;
 use Sentry\SentryBundle\EventListener\MessengerListener;
 use Sentry\SentryBundle\SentryBundle;
@@ -73,7 +74,7 @@ final class SentryExtension extends ConfigurableExtension
     {
         $options = $config['options'];
 
-        if (isset($config['dsn'])) {
+        if (\array_key_exists('dsn', $config)) {
             $options['dsn'] = $config['dsn'];
         }
 
@@ -140,6 +141,8 @@ final class SentryExtension extends ConfigurableExtension
         if (!$config['register_error_listener']) {
             $container->removeDefinition(ErrorListener::class);
         }
+
+        $container->getDefinition(ConsoleListener::class)->setArgument(1, $config['register_error_listener']);
     }
 
     /**